+++ crates/objc2/src/__macro_helpers/defined_ivars.rs
@@ -234,35 +234,34 @@ where
     }
 }
 
-/// Register the class, and get the ivar offsets.
 #[inline]
-pub(crate) fn register_with_ivars<T: DefinedClass>(
-    mut builder: ClassBuilder,
-) -> (&'static AnyClass, isize, isize) {
-    let (ivar_name, drop_flag_name): (Cow<'static, CStr>, Cow<'static, CStr>) = {
-        if cfg!(feature = "gnustep-1-7") {
-            // GNUStep does not support a subclass having an ivar with the
-            // same name as a superclass, so let's use the class name as the
-            // ivar name to ensure uniqueness.
+pub(crate) fn ivar_drop_flag_names<T: DefinedClass>() -> (Cow<'static, CStr>, Cow<'static, CStr>) {
+    if cfg!(feature = "gnustep-1-7") {
+        // GNUStep does not support a subclass having an ivar with the
+        // same name as a superclass, so let's use the class name as the
+        // ivar name to ensure uniqueness.
+        (
+            CString::new(format!("{}_ivars", T::NAME)).unwrap().into(),
+            CString::new(format!("{}_drop_flag", T::NAME))
+                .unwrap()
+                .into(),
+        )
+    } else {
+        // SAFETY: The byte slices are NUL-terminated, and do not contain
+        // interior NUL bytes.
+        // TODO: Use `c"my_str"` syntax once in MSRV
+        unsafe {
             (
-                CString::new(format!("{}_ivars", T::NAME)).unwrap().into(),
-                CString::new(format!("{}_drop_flag", T::NAME))
-                    .unwrap()
-                    .into(),
+                CStr::from_bytes_with_nul_unchecked(b"ivars\0").into(),
+                CStr::from_bytes_with_nul_unchecked(b"drop_flag\0").into(),
             )
-        } else {
-            // SAFETY: The byte slices are NUL-terminated, and do not contain
-            // interior NUL bytes.
-            // TODO: Use `c"my_str"` syntax once in MSRV
-            unsafe {
-                (
-                    CStr::from_bytes_with_nul_unchecked(b"ivars\0").into(),
-                    CStr::from_bytes_with_nul_unchecked(b"drop_flag\0").into(),
-                )
-            }
         }
-    };
+    }
+}
 
+/// Register the ivars.
+#[inline]
+pub(crate) fn register_ivars<T: DefinedClass>(builder: &mut ClassBuilder, ivars_name: &CStr) {
     if T::HAS_IVARS {
         // TODO: Consider not adding a encoding - Swift doesn't do it.
         let ivar_encoding = Encoding::Array(
@@ -279,25 +278,33 @@ pub(crate) fn register_with_ivars<T: DefinedClass>(
                 _ => &Encoding::None,
             },
         );
-        unsafe { builder.add_ivar_inner::<T::Ivars>(&ivar_name, &ivar_encoding) };
+        unsafe { builder.add_ivar_inner::<T::Ivars>(ivars_name, &ivar_encoding) };
     }
+}
 
+/// Register the drop flag ivar.
+#[inline]
+pub(crate) fn register_drop_flag<T: DefinedClass>(
+    builder: &mut ClassBuilder,
+    drop_flag_name: &CStr,
+) {
     if T::HAS_DROP_FLAG {
         // TODO: Maybe we can reuse the drop flag when subclassing an already
         // defined class?
-        builder.add_ivar::<DropFlag>(&drop_flag_name);
+        builder.add_ivar::<DropFlag>(drop_flag_name);
     }
+}
 
-    let cls = builder.register();
-
-    let ivars_offset = if T::HAS_IVARS {
+#[inline]
+pub(crate) fn ivars_offset<T: DefinedClass>(cls: &AnyClass, ivars_name: &CStr) -> isize {
+    if T::HAS_IVARS {
         // Monomorphized error handling
         // Intentionally not #[track_caller], we expect this error to never occur
         fn get_ivar_failed() -> ! {
             unreachable!("failed retrieving instance variable on newly defined class")
         }
 
-        cls.instance_variable(&ivar_name)
+        cls.instance_variable(ivars_name)
             .unwrap_or_else(|| get_ivar_failed())
             .offset()
     } else {
@@ -306,16 +313,19 @@ pub(crate) fn register_with_ivars<T: DefinedClass>(
         // This is fine, since any reads here will only be via zero-sized
         // ivars, where the actual pointer doesn't matter.
         0
-    };
+    }
+}
 
-    let drop_flag_offset = if T::HAS_DROP_FLAG {
+#[inline]
+pub(crate) fn drop_flag_offset<T: DefinedClass>(cls: &AnyClass, drop_flag_name: &CStr) -> isize {
+    if T::HAS_DROP_FLAG {
         // Monomorphized error handling
         // Intentionally not #[track_caller], we expect this error to never occur
         fn get_drop_flag_failed() -> ! {
             unreachable!("failed retrieving drop flag instance variable on newly defined class")
         }
 
-        cls.instance_variable(&drop_flag_name)
+        cls.instance_variable(drop_flag_name)
             .unwrap_or_else(|| get_drop_flag_failed())
             .offset()
     } else {
@@ -324,9 +334,7 @@ pub(crate) fn register_with_ivars<T: DefinedClass>(
         // This is fine, since the drop flag is never actually used in the
         // cases where it was not added.
         0
-    };
-
-    (cls, ivars_offset, drop_flag_offset)
+    }
 }
 
 /// # Safety