+++ crates/objc2/src/__macro_helpers/mod.rs
@@ -3,6 +3,7 @@ pub use core::cell::UnsafeCell;
 pub use core::cmp::{Eq, PartialEq};
 pub use core::convert::AsRef;
 pub use core::default::Default;
+pub use core::ffi::CStr;
 pub use core::fmt;
 pub use core::hash::{Hash, Hasher};
 pub use core::marker::{PhantomData, Sized};
@@ -35,8 +36,8 @@ pub use self::class::{DoesNotImplDrop, MainThreadOnlyDoesNotImplSendSync, ValidT
 pub use self::common_selectors::{alloc_sel, dealloc_sel, init_sel, new_sel};
 pub use self::convert::{ConvertArgument, ConvertArguments, ConvertReturn, TupleExtender};
 pub use self::define_class::{
-    ClassBuilderHelper, ClassProtocolMethodsBuilder, MaybeOptionRetained, MessageReceiveRetained,
-    RetainedReturnValue, ThreadKindAutoTraits,
+    class_c_name, define_class, ClassBuilderHelper, ClassProtocolMethodsBuilder,
+    MaybeOptionRetained, MessageReceiveRetained, RetainedReturnValue, ThreadKindAutoTraits,
 };
 pub use self::defined_ivars::DefinedIvarsHelper;
 pub use self::image_info::ImageInfo;