+++ crates/objc2/CHANGELOG.md
@@ -7,6 +7,13 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ## Unreleased - YYYY-MM-DD
 
 
+## [0.6.3] - 2025-10-04
+[0.6.3]: https://github.com/madsmtm/objc2/compare/objc2-0.6.2...objc2-0.6.3
+
+## Fixed
+* Fixed documentation on docs.rs.
+
+
 ## [0.6.2] - 2025-08-14
 [0.6.2]: https://github.com/madsmtm/objc2/compare/objc2-0.6.1...objc2-0.6.2